W 7198137 no eslint for aura (#69)
* @W-7198137

eslint should not be run on aura components

Includes a change to an aura component to confirm the ignore is working.  Will then revert the test change

https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000079iQgIAI/view

* @W-7198137

eslint should not be run on aura components

Includes a change to an aura component to confirm the ignore is working.  Will then revert the test change

https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000079iQgIAI/view

* @W-7198137

eslint should not be run on aura components

Includes a change to an aura component to confirm the ignore is working.  Will then revert the test change

https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000079iQgIAI/view