W-9894880 - Adding enable-animated-inputs flag
On of the most common tasks all people do when customizing newport is disable the animated labels. Unfortunately it's not a straightforward task. So this PR is making it straight forward by adding a simple boolean config property to turn them on or off.

I need this for the OS Zen Garden css files because in the ones I'm creating them all need the animated labels turned off. So doing this saves a lot of time.

In addition, the way the webpack generation was happening, it was incredibly slow because it had to rebuild the entire JS React app if a css change was made due to the CSS being included in preview.js. So I've sped it up by adding some simple JS into the preview-head.html which listens for a rebuild command and updates the css that way. The additional benefit of this is that this css also includes SourceMaps so chrome and other browsers will now nicely link back to the source file for specific files!

Also tidied up commented out dead code.