W-9894880 - config option to render bordered input
A common initial change after removing the animated labels is to get back to bordered slds style input boxes. This is another config switch I've added to support this.

Since bordered inputs never make sense with the animated logo, it's controlled by the animated logo property too. So both have to be set to false for it to work.