W-9894880 - Adding docs page on css zen garden
This page talks about how to get started with the css zengarden demo. It covers how to checkout the branch, include the OS we rebranded as a datapack, the steps taken to rebrand it.

I've also fixed up styling of code samples. Also for some reason the docs had been removed from the previewer so added them back in.