Only set error on assignee picker if it's present @W-9560156@ (#2244)
* @W-9560156@
@https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07AH000000QPPrYAO/view@
@rev rsrikantan@

Only set error to the assignee picker if it is present.  Otherwise, we trust the picker to maintain its own error

* @W-9560156@
@https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07AH000000QPPrYAO/view@
@rev rsrikantan@

Only set error to the assignee picker if it is present.  Otherwise, we trust the picker to maintain its own error

* @W-9560156@
@https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07AH000000QPPrYAO/view@
@rev rsrikantan@

Only set error to the assignee picker if it is present.  Otherwise, we trust the picker to maintain its own error