@W-11206393 Fix stack overflow in apexClassHasSomePropertyMatching. (#2785)
* @W-11206393 Fix stack overflow in apexClassHasSomePropertyMatching.

apexClassHasSomePropertyMatching recursively goes through the properties
of an Apex class and checks if it matches a given filter. Since the
function doesn't keep track of which Apex classes it already seen, it
can get stuck infinitely recursing over them if they happen to refer to
each other. eg. A -> A or A -> B -> A.

This change just uses a simple hash to mark which Apex classes are
currently being processed.

* make it pretty

* use recursive helper and add jsdocs

* just pass hash

* only check for duplicate apex classes